requirements.txt
pdfplumber
pandas
pypdf
pdf2image
pytesseract
Pillow
